the two spellchecker and preprocessing files are merged - They still have work to do at this point
